refactor: tablegroup 도메인과 table 도메인 통합
