+++ src/main/java/kitchenpos/table/domain/OrderTableRepository.java
@@ -1,6 +1,5 @@
 package kitchenpos.table.domain;
 
-import kitchenpos.tablegroup.domain.TableGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
