+++ src/main/java/kitchenpos/table/domain/TableGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.tablegroup.domain;
+package kitchenpos.table.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
