+++ src/test/java/kitchenpos/table/domain/OrderTableTest.java
@@ -1,6 +1,5 @@
 package kitchenpos.table.domain;
 
-import kitchenpos.tablegroup.domain.TableGroup;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
